requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dm-tree==0.1.8
environs==14.1.1
fastapi==0.115.11
fastapi_cors==0.0.6
greenlet==3.1.1
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
keras==3.9.0
keras-core==0.1.7
markdown-it-py==3.0.0
marshmallow==3.26.1
mdurl==0.1.2
ml_dtypes==0.5.1
mysql-connector-python==9.2.0
namex==0.0.8
numpy==2.0.2
optree==0.14.1
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
rich==13.9.4
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.2
